#227 Decompose the SqlRepository god class into focused collaborators
Split the 1139-line SqlRepository into four internal collaborators — SqlSchemaInstaller
(FK/DDL/junction install), SqlEntityLoader (bulk load + junction hydration), SqlWritePipeline
(insert/update/delete single+batch), OptimisticLockRecovery (conflict recovery + stale-id
draining) — leaving SqlRepository a thin orchestrator with a byte-identical public surface.

Retains fail-loud co-presence guards in init: JunctionAware descriptor/accessor agreement and
@Version-column/VersionedTableDef pairing, which the capability-interface split alone cannot
enforce.
